+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 * 0.1.1 (2023-06-17): Add changelog, fix readme
 * 0.2.0 (2023-06-17): Add `--sloccount` option; update dependencies
 * 0.2.1 (2023-06-17): Fix typo
+* 0.3.0 (2023-06-18): Expose as library; convert PARAMS static HashMap into
+  ProjectType enum and remove `lazy_static` dependency; add `--exclude` option;
+  update dependencies